수정 : schemas 구조 변경 및 예시 추가
