+++ app/schemas/__init__.py
@@ -1,10 +1,16 @@
-from app.schemas.user        import UserBase, UserCreate, UserUpdate, UserResponse
 from app.schemas.responses   import (
                                 SuccessResponseBase,
                                 ErrorResponseBase,
                                 MethodNotAllowedHandling
                                 )
-from app.schemas.users       import NotFoundUserHandling
+from app.schemas.users       import (
+                                NotFoundUserHandling,
+                                UnauthorizedHandler,
+                                UserBase,
+                                UserCreate,
+                                UserUpdate,
+                                UserResponse
+                                )
 from app.schemas.study_rooms import (
                                 StudyRoomsCreate,
                                 StudyRoomsUpdate,
@@ -15,4 +21,3 @@ from app.schemas.study_rooms import (
                                 BodyNeedyStudyRoomHandling,
                                 QueryNeedyStudyRoomHandling
                                 )
-
