+++ app/schemas/users/__init__.py
@@ -1,3 +1,10 @@
 from app.schemas.users.handling import (
-                                    NotFoundUserHandling
+                                    NotFoundUserHandling,
+                                    UnauthorizedHandler
+                                    )
+from app.schemas.users.crud        import (
+                                    UserBase,
+                                    UserCreate,
+                                    UserUpdate,
+                                    UserResponse
                                     )
